New GUI: working on the simulation panel

+++ www/phensim/app/Http/Livewire/Simulations/DataTable.php
@@ -15,16 +15,17 @@ class DataTable extends Component
     use WithPagination;
 
     protected $paginationTheme = 'bootstrap';
-    protected $listeners = ['deleteSimulation'];
+    protected $listeners = ['receivedConfirmation'];
 
-    public $categories = [];
-    public $sortColumn = 'name';
-    public $sortDirection = 'asc';
+    public $sortColumn = 'created_at';
+    public $sortDirection = 'desc';
     public $perPage = 10;
     public $searchColumns = [
         'name'   => '',
         'status' => -1,
     ];
+    public $displayingLog = false;
+    public $currentSimulationId = null;
 
     public function sortByColumn($column): void
     {
@@ -63,6 +64,12 @@ class DataTable extends Component
         return $simulations;
     }
 
+    public function displayLogs(int $simulationId): void
+    {
+        $this->displayingLog = true;
+        $this->currentSimulationId = $simulationId;
+    }
+
     /**
      * Confirm that the given simulation should be deleted.
      *
@@ -73,11 +80,54 @@ class DataTable extends Component
     public function confirmSimulationDeletion(int $simulationId): void
     {
         $this->dispatchBrowserEvent(
-            'swal:confirm:delete',
+            'swal:confirm',
             [
-                'type'  => 'warning',
+                'type'  => 'delete',
+                'icon'  => 'warning',
                 'title' => __('Delete Simulation'),
                 'text'  => __('Are you sure you would like to delete this simulation?'),
+                'id'    => $simulationId,
+            ]
+        );
+    }
+
+    /**
+     * Confirm that the given simulation should be submitted.
+     *
+     * @param  int  $simulationId
+     *
+     * @return void
+     */
+    public function confirmSimulationSubmission(int $simulationId): void
+    {
+        $this->dispatchBrowserEvent(
+            'swal:confirm',
+            [
+                'type'  => 'submit',
+                'icon'  => 'info',
+                'title' => __('Submit Simulation'),
+                'text'  => __('Are you sure you wish to submit this simulation?'),
+                'id'    => $simulationId,
+            ]
+        );
+    }
+
+    /**
+     * Confirm that the given simulation should be resubmitted.
+     *
+     * @param  int  $simulationId
+     *
+     * @return void
+     */
+    public function confirmSimulationReSubmission(int $simulationId): void
+    {
+        $this->dispatchBrowserEvent(
+            'swal:confirm',
+            [
+                'type'  => 'resubmit',
+                'icon'  => 'info',
+                'title' => __('Submit Simulation'),
+                'text'  => __('Are you sure you wish to submit this simulation?'),
                 'id'    => $simulationId,
             ]
         );
@@ -87,14 +137,20 @@ class DataTable extends Component
      * Delete the simulation.
      *
      * @param  \App\Models\Simulation  $simulation
+     * @param  string  $type
      *
      * @return void
      * @throws \Exception
      */
-    public function deleteSimulation(Simulation $simulation): void
+    public function receivedConfirmation(Simulation $simulation, string $type): void
     {
-        if ($simulation->canBeDeleted()) {
+        if ($type === 'delete' && $simulation->canBeDeleted()) {
+            $simulation->deleteJobDirectory();
             $simulation->delete();
+        } elseif ($type === 'submit' && $simulation->isReady()) {
+            $simulation->submit();
+        } elseif ($type === 'resubmit' && $simulation->isFailed()) {
+            $simulation->reSubmit();
         }
     }
 
@@ -105,7 +161,8 @@ class DataTable extends Component
         return view(
             'livewire.simulations.data-table',
             [
-                'simulations' => $simulations->paginate($this->perPage),
+                'simulations'       => $simulations->paginate($this->perPage),
+                'currentSimulation' => ($this->displayingLog) ? Simulation::where('id', $this->currentSimulationId)->firstOrFail() : null,
             ]
         );
     }
